Change role id from serial to bigserial Change column type in user_role table from integer to bigint

+++ src/main/resources/db/migration/V2__add_user_role_entities.sql
@@ -1,6 +1,6 @@
 CREATE TABLE roles
 (
-    id serial PRIMARY KEY,
+    id   bigserial PRIMARY KEY,
     name character varying(255) NOT NULL
 );
 
@@ -16,7 +16,7 @@ CREATE TABLE users
 CREATE TABLE user_role
 (
     user_id uuid NOT NULL references users (uuid),
-    role_id   integer NOT NULL references roles (id)
+    role_id   bigint NOT NULL references roles (id)
 );
 
 CREATE INDEX users_uuid ON users (uuid);
